+++ api/maven-api-core/src/main/java/org/apache/maven/api/services/ChecksumAlgorithmService.java
@@ -19,6 +19,7 @@
 package org.apache.maven.api.services;
 
 import java.io.*;
+import java.nio.ByteBuffer;
 import java.nio.file.Path;
 import java.util.Collection;
 import java.util.Map;
@@ -84,4 +85,48 @@ public interface ChecksumAlgorithmService extends Service {
     @Nonnull
     Map<String, String> calculate(@Nonnull Path file, @Nonnull Collection<ChecksumAlgorithm> algorithms)
             throws IOException;
+
+    /**
+     * The checksum algorithm.
+     */
+    interface ChecksumAlgorithm {
+        /**
+         * Returns the algorithm name, usually used as key, never {@code null} value. The name is a standard name of
+         * algorithm (if applicable) or any other designator that is algorithm commonly referred with. Example: "SHA-1".
+         */
+        @Nonnull
+        String getName();
+
+        /**
+         * Returns the file extension to be used for given checksum file (without leading dot), never {@code null}. The
+         * extension should be file and URL path friendly, and may differ from algorithm name.
+         * The checksum extension SHOULD NOT contain dot (".") character.
+         * Example: "sha1".
+         */
+        @Nonnull
+        String getFileExtension();
+
+        /**
+         * Each invocation of this method returns a new instance of calculator, never {@code null} value.
+         */
+        @Nonnull
+        ChecksumCalculator getCalculator();
+    }
+
+    /**
+     * The checksum calculator.
+     */
+    interface ChecksumCalculator {
+        /**
+         * Updates the checksum algorithm inner state with input.
+         */
+        void update(ByteBuffer input);
+
+        /**
+         * Returns the algorithm end result as string, never {@code null}. After invoking this method, this instance should
+         * be discarded and not reused. For new checksum calculation you have to get new instance.
+         */
+        @Nonnull
+        String checksum();
+    }
 }
